requirements.txt
langchain==0.0.141
openai
faiss-cpu
streamlit
streamlit-chat
unstructured